Ensure that parents can be added with an explicit declaration
Newer factory boy versions require an explicit declaration on the
model. As the parent did not previously exists as a declaration it
would error out.

Moving it to a post gen works, but the classes parent factory needs
to come along for the ride. Also removes the explicit need for a root
node, due to technical limitations `extracted_parent` is `None` when
no `parent` variable is passed, instead assumes you implicitly want
root unless otherwise stated.