Add handling for deep declarations to StructBlockFactory
Prior to this commit, it was not possible to use deep declarations on
a StructBlockFactory for a struct block that contains a stream
block. factory_boy doesn't handle deep declarations with indices,
e.g. struct_block__inner_stream__0__struct_block, so when the work is
deferred to factory.BaseFactory._generate via StructBlockFactory we
get an error.

Unfortunately this means we can't rely on factory_boy's handling for
generating nested objects, hence the amount of logic on
StructBlockFactory.